src/requirements.txt
attrs==21.3.0
black==21.12b0
dateparser==1.1.0
psycopg2-binary==2.9.2
python-dotenv==0.19.2
requests==2.26.0
SQLAlchemy==1.4.29
